Imported Upstream version 2.4.4

+++ imap/saslclient.c
@@ -39,7 +39,7 @@
  * AN ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING
  * OUT OF OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
  *
- * $Id: saslclient.c,v 1.19 2008/10/08 15:47:08 murch Exp $
+ * $Id: saslclient.c,v 1.20 2010/01/06 17:01:39 murch Exp $
  */
 
 #include <config.h>
@@ -227,7 +227,7 @@ int saslclient(sasl_conn_t *conn, struct sasl_cmd_t *sasl_cmd,
 	clientout = NULL;
     }
     else if (!sendliteral &&
-	     ((int) (strlen(cmdbuf) + clientoutlen + 3) > sasl_cmd->maxlen)) {
+	     ((strlen(cmdbuf) + clientoutlen + 3) > sasl_cmd->maxlen)) {
 	/* initial response is too long for auth command,
 	   so wait for a server challenge before sending it */
 	goto noinitresp;
